Build PIC versions of libafsauthent and libafsrpc
For Debian, we don't ship the shared libraries due to concerns over the
stability of the interface.  It's useful, however, to have a PIC library
for consumers of libopenafs-dev who want to link AFS code into shared
libraries.  Generate static PIC versions of the two libraries that are
already built PIC for that purpose.