linux-fh-based-cache-20090511
LICENSE IPL10
FIXES 123620

rework previous fh-based cache support. now allows use of
arbitrarily large file handles, and build correctly for UKERNEL.

we now just use this if there's no iget, instead of providing a configure
switch.

because this relies on the somewhat invasive patch from 124184 it is not
currently pulled up to 1.4.x