DEVEL15-linux26-defer-cred-changing-20090511
LICENSE IPL10
FIXES 124737

Newer Linux kernels differentiate between the real and effective
credentials of a process, and prevent a process's credentials from being
change when the effective credentials have been set to a different value.
When AFS notices a keyring PAG exists but no group-based PAG does, the
attempt to rectify this, if done in a VFS call (which changes effective creds)
triggers this issue. We defer the change to the groups to avoid it.


(cherry picked from commit 7d530b9080f6dd4d7f7e4555ed6ce855ff81bcaa)