DEVEL15-man-page-cellservdb-updates-20090518
FIXES 124794
LICENSE IPL10

Note in CellServDB man page that it's also used to populate root.afs for
a -dynroot client.  Also document the dynamic lookup of database servers
with -afsdb and provide some more information about when CellServDB has to
contain the cell and when it doesn't.

Mark the backup server as optional, and indicate that the authentication
server is deprecated and CellServDB isn't required for authentication if
Kerberos v5 and aklog are used.


(cherry picked from commit 6df60468130977744fea0499fba2351ef74175d8)