doc-xml-build-20090518
LICENSE MIT

Standardize the UNIX Makefiles for all of the DocBook guides.  Remove the
rest of the generated files and switch to xsltproc and dblatex for the
document generation in all cases.  Fix a few DocBook errors by removing
the contents of the <index> tag and removing the unknown <pubsnumber> tag
in the <revision> field.