windows-btree-flush-race-20090522
LICENSE MIT
FIXES 124787

A flush operation on a directory will result in the btree being destroyed.
This can race with on-going operations.  Make sure that the dirlock is held
if the btree is destroyed.  Otherwise, just invalidate the btree version
number.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
LICENSE MIT
FIXES 124787

correct sandbox leakage