DEVEL15-windows-dirty-buffers-20090608
LICENSE MIT

In the buf_IncrSync thread, if the volume is known to be unavailable
do not attempt to write the buffer.  Just skip it.

In buf_Sync(), if we are shutting down and a buffer is left dirty,
log a message to the Windows Event Log indicating that the dirty buffer
was lost.


(cherry picked from commit 2033fa9ec62089887f074381ac118a2c60e9c311)