windows-callback-server-tracking-for-readonly-vols-20090626
LICENSE MIT

Readonly volume callbacks apply to every cm_scache_t from the
same volume.  We were already tracking the cbExpires time in
the cm_volume_t.  We need to also track the cbServerp in the
cm_volume_t so that the cm_scache_t objects can be assigned
the correct server from which the callback was issued.

Add cbServerpRO field to the cm_volume_t.  Bump the memory
map version to force a rebuild of the cache data.