DEVEL15-windows-multi-homed-callbacks-20090627
LICENSE MIT

Properly handle callbacks from multi-homed file servers.
Comparing cm_server_t pointers is insufficient.  For a multi-homed
server there will be multiple entries.  The UUID for all of the
equivalent entries will be the same.  What matters is not that
the pointers are the same but whether in the case of UUID labeled
servers that the UUIDs match.

Add cm_ServerEqual() to perform the comparison.


(cherry picked from commit d54859c6ac2e5186b18da027303860966504dbd9)