Use InterlockedExchangeAdd() instead of InterlockedAdd() on Windows
InterlockedAdd() is not supported on all the platforms we wish to
support.  However, InterlockedExchangeAdd() is.  The two
functions are semantically similar with the exception that
InterlockedExchangeAdd() returns the initial value of the object
being modified while InterlockedAdd() returns the resulting
value.  Since the return value of the rx_MutexAdd() macro is
unused, this is irrelevant.

Reviewed-on: http://gerrit.openafs.org/98
Verified-by: Jeffrey Altman <jaltman@openafs.org>
Reviewed-by: Jeffrey Altman <jaltman@openafs.org>