Refactor OpenBSD param files into common and architecture-specific sections.
The new structure follows the pattern of the NetBSD param files, as
suggested by Derrick.  Since the change is retroactive, it seems
reasonable to also make the retroactive change to add the amd64 param
files, back to OpenBSD 3.6.  OpenBSD 3.6 added SMP support for x86_64,
so the architecture was a mature subtarget at that point.

Reviewed-on: http://gerrit.openafs.org/http://gerrit.openafs.org/177
Tested-by: Derrick Brashear <shadow@dementia.org>
Reviewed-by: Derrick Brashear <shadow@dementia.org>