Build fixes: prevent unnecessary rebuilding
Fixes to a few Makefiles to prevent needless rebuilding when using make
after a successful build.

Rework dependencies in src/bubasics to prevent needless rebuilding and
fix some remaining parallel build issues.  Multiple targets really
have to be avoided.

Remove dependency on the include/afs directory from the src/config Makefile.
Depending on a directory means that we will rebuild the target if the
timestamp changes, for instance if other files or sub-directories are
installed.  This caused param.h to get re-installed in some cases and
the kernel module source files to be recompiled.  This directory is
created in the top-level Makefile anyway so we shouldn't need an
explicit dependency.

Reviewed-on: http://gerrit.openafs.org/http://gerrit.openafs.org/171
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>