Finish afscell implementation work
This patch completes the implementation of an install time prompeter
for information about the local AFS cell, which is then used to populate
the ThisCell file. It removes the need for manual editing of
configuration files as part of the installation.

(Build system modified by Simon Wilkinson from Toby's original patch)

Reviewed-on: http://gerrit.openafs.org/http://gerrit.openafs.org/201
Tested-by: Derrick Brashear <shadow@dementia.org>
Reviewed-by: Derrick Brashear <shadow@dementia.org>