Running bos -noauth should not require client CellServDB configuration
When setting up a new cell, the admin is instructed to use "bos -noauth"
to configure the servers and only afterwards create the client configuration.
This leads to a catch-22 since currently bos -noauth fails if there is no
client configuration even though it is not required.

This change avoids the call to afsconf_Open() which is only required
if tokens are being acquired.

FIXES 125156

Reviewed-on: http://gerrit.openafs.org/209
Reviewed-by: Simon Wilkinson <sxw@inf.ed.ac.uk>
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Jeffrey Altman <jaltman@openafs.org>
Reviewed-by: Jeffrey Altman <jaltman@openafs.org>
(cherry picked from commit e435e248f44c46866cb73db185b9548c0bd148b4)
Change-Id: I5a40091410bcc4b28fce04b7cd65ae797b3d8ed0
Reviewed-on: http://gerrit.openafs.org/942
Tested-by: Derrick Brashear <shadow@dementia.org>