Windows: Preserve GetLastError during WinTorture testing
When an error occurs during the WinTorture scripts the
actual error was not always being preserved because
subsequent Win32 calls were executed before the value
was logged.

In at least one case, the variable used to store the
last error was not initialized nor assigned in all data
paths.

LICENSE MIT

Reviewed-on: http://gerrit.openafs.org/270
Tested-by: Jeffrey Altman <jaltman@openafs.org>
Reviewed-by: Jeffrey Altman <jaltman@openafs.org>