Windows: Setting Server Preferences
The Windows cache manager can apply administrator specified
server preferences as specified in the registry.   When these
rankings are applied the CM_SERVERFLAG_PREF_SET flag was not
set on the cm_server_t object.  In addition, appropriate locking
was not being used in the places where the flag was set.

LICENSE MIT

Reviewed-on: http://gerrit.openafs.org/303
Tested-by: Jeffrey Altman <jaltman@openafs.org>
Reviewed-by: Jeffrey Altman <jaltman@openafs.org>