Handle more RPC service names
While the Windows client only implements support for WKSSVC and
SRVSVC RPC services, it should be aware of and handle requests
for other well known service names.

This patch refactors service name checks to a separate
MSRPC_IsWellKnownService() function.  In addition, it separates
the tests for IPC TIDs from the test for well known service
names.  An IPC TID may be used for generic pipe operations as
well as for RPC message pipe operations.  Therefore the fact that
a TID was opened for IPC doesn't necessarily mean that it will
only be used for RPC.

Reviewed-on: http://gerrit.openafs.org/307
Tested-by: Jeffrey Altman <jaltman@openafs.org>
Reviewed-by: Jeffrey Altman <jaltman@openafs.org>