Windows: Prevent smb_StartedLock use before initialization
The smb_StartedLock osi_mutex can be obtained before
initialization due to a race if the IP address configuration
is in a state of flux as afsd_service is starting up.
To avoid the problem call cm_InitDaemons() after smb_Init().

This race is the cause of the periodic assertions in some
environments when lock order validation is enabled.

LICENSE MIT

Reviewed-on: http://gerrit.openafs.org/318
Tested-by: Jeffrey Altman <jaltman@openafs.org>
Reviewed-by: Jeffrey Altman <jaltman@openafs.org>