Windows: cm_conn_t must not be referenced after cm_Analyze is called
The 64-bit Large File Support added a reference to the cm_conn_t
object after cm_Analyze() is called and releases the hold.  This
permits the object to be freed if it is appropriate to do so
which can result in an invalid memory reference.

This change removes the improper reference.

LICENSE MIT

Reviewed-on: http://gerrit.openafs.org/326
Tested-by: Asanka Herath <asanka@secure-endpoints.com>
Reviewed-by: Asanka Herath <asanka@secure-endpoints.com>
Tested-by: Jeffrey Altman <jaltman@openafs.org>
Reviewed-by: Jeffrey Altman <jaltman@openafs.org>