Windows: cleanup smb polution of non-smb modules
There are a variety of functions that are declared in smb modules
that are used within non-smb modules.   In order to reduce some of
the polution, several time conversion functions and the thread
priority operations are migrated to cm_utils.  As part of this
migration the unused !USE_NUMERIC_TIME_CONV time implementations
are removed.

In addition, an unnecessary use of an SMB preprocessor symbols
is removed from cm_freelance.c and smb_ioctls.h is no longer
included within afsd.h.

LICENSE MIT

Reviewed-on: http://gerrit.openafs.org/330
Reviewed-by: Jeffrey Altman <jaltman@openafs.org>
Tested-by: Jeffrey Altman <jaltman@openafs.org>