Only force a specific compiler where necessary
Rather than setting the compiler to cc on all platforms by default,
use the Autoconf default unless the specific compiler might matter.
Conservatively set the compiler to cc for all non-Linux, non-BSD
operating systems that previously didn't explicitly set it.  Fix a
few places on Linux where the compiler was set to gcc explicitly
rather than using the Autoconf result.

Don't bother setting the Linux i386 compiler to gcc -pipe.  The compile
performance difference isn't enough to bother with the special
exception, and we don't do this for other Linux systems.

With this change, setting CC at Autoconf time will correctly set the
compiler per the user's preferences unless they're on an operating
system where we care about having a particular compiler for kernel
purposes.  This can be relaxed further in the future, particularly if
the kernel compiler is broken out to a separate setting.

Reviewed-on: http://gerrit.openafs.org/337
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>